Initial update to 1.8, Super beta. Most rendering related hooks are out due to major changes in 1.8.
Some notes:
Almost all int x, int y, int z parameters have been changed to BlockPos class
ForgeDirection has been removed, replaced by net.minecraft.util.EnumFacing.
All FML classes have moved from packet cpw.mods.fml to net.minecraftforge.fml
Fluid Rendering has been disabled for the time being, to be re-evaulated and a test mod created for it.
Minecraft now uses a Model based system for rendering blocks and Items. The intention is to expand the model format to better suit modder's needed once it is evaulated.
As such, The model loaders from Forge have been removed, to be replaced by expanding vanilla's model format.
Metadata has been extracted out in Minecraft to IBlockState, which holds a list of properties instead of magic number metadata. DO NOT listen to the fearmongering, you can do EVERYTHING with block states you could previously with metadata.
Stencil Bits are disabled entirely by for the main Display, Modders must enable and recreate the FrameBuffer if they wish to use Stencil Bits.

+++ src/main/java/net/minecraftforge/event/terraingen/WorldTypeEvent.java
@@ -1,6 +1,6 @@
 package net.minecraftforge.event.terraingen;
 
-import cpw.mods.fml.common.eventhandler.Event;
+import net.minecraftforge.fml.common.eventhandler.Event;
 import net.minecraft.world.gen.layer.GenLayer;
 import net.minecraft.world.WorldType;
 
@@ -39,10 +39,10 @@ public class WorldTypeEvent extends Event
      **/
     public static class BiomeSize extends WorldTypeEvent
     {
-        public final byte originalSize;
-        public byte newSize;
-        
-        public BiomeSize(WorldType worldType, byte original)
+        public final int originalSize;
+        public int newSize;
+
+        public BiomeSize(WorldType worldType, int original)
         {
             super(worldType);
             originalSize = original;
@@ -52,7 +52,7 @@ public class WorldTypeEvent extends Event
 
     /**
      * InitBiomeGens is fired when vanilla Minecraft attempts to initialize the biome generators.<br>
-     * This event is fired just during biome generator initialization in 
+     * This event is fired just during biome generator initialization in
      * WorldChunkManager#WorldChunkManager(long, WorldType). <br>
      * <br>
      * {@link #seed} the seed of the world. <br>
@@ -71,7 +71,7 @@ public class WorldTypeEvent extends Event
         public final long seed;
         public final GenLayer[] originalBiomeGens;
         public GenLayer[] newBiomeGens;
-        
+
         public InitBiomeGens(WorldType worldType, long seed, GenLayer[] original)
         {
             super(worldType);
